requirements.txt
streamlit
pandas
folium
scikit-learn
numpy
scipy
pymoo
numba==0.57
biopython
peptides